printing something near to a fractal and with a debug printf in the code

+++ inc/fractol.h
@@ -6,7 +6,7 @@
 /*   By: davifern <davifern@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/08/28 15:50:36 by davifern          #+#    #+#             */
-/*   Updated: 2023/09/18 16:58:19 by davifern         ###   ########.fr       */
+/*   Updated: 2023/09/19 21:16:24 by davifern         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -34,6 +34,6 @@ typedef struct s_img
 
 typedef struct s_complex
 {
-	int	x;
-	int	y;
+	double	x;
+	double	y;
 }	t_complex;
